xl: Avoid called getdents for folders in listing
